+++ apps/mobile/src/components/posts/link/index.tsx
@@ -120,7 +120,7 @@ export function PostLinkCard({
 
 const stylesheet = createStyleSheet((theme) => ({
   compact: (large?: boolean) => ({
-    backgroundColor: theme.colors.gray.ui,
+    backgroundColor: theme.colors.gray.uiActive,
     borderCurve: 'continuous',
     borderRadius: theme.space[large ? 2 : 1],
     height: theme.space[8] * (large ? 2 : 1),