+++ apps/mobile/src/components/posts/video/player.tsx
@@ -218,7 +218,7 @@ const stylesheet = createStyleSheet((theme, runtime) => ({
     justifyContent: 'center',
   },
   compact: (large?: boolean) => ({
-    backgroundColor: theme.colors.gray.ui,
+    backgroundColor: theme.colors.gray.uiActive,
     borderCurve: 'continuous',
     borderRadius: theme.space[large ? 2 : 1],
     height: theme.space[8] * (large ? 2 : 1),